Got e2fsck linking. Compiling and linking various ext2fs libraries. The kernel compiles under 32-bit mode.

+++ src/gnu/ext2fs/i386-bitops.h
@@ -144,7 +144,7 @@ static __inline__ unsigned long ffz(unsigned long word)
 /*
  * find the first occurrence of byte 'c', or 1 past the area if none
  */
-static __inline__ char * memscan(void * addr, unsigned char c, int size)
+static __inline__ char * memscan(void * addr, unsigned char c, size_t size)
 {
         if (!size)
                 return addr;
